Refactor: AppleSignInHelper의 Nonce 생성하는 코드를 별도 파일로 분리
CryptoUtils는 매 번 새 nonce가 생성 되어야 하므로 class 대신 struct로 생성하였다.
AppleSignInHelper가 호출되면 nonce도 생성된다.
