+++ PJ3T3_Postie/Manager/AppleSignInHelper.swift
@@ -9,13 +9,13 @@ import AuthenticationServices
 import CryptoKit
 
 final class AppleSignInHelper: NSObject, ObservableObject {
+    private var cryptoUtils = CryptoUtils()
     private var nonce = ""
     
-    
     func signInWithAppleRequest(_ request: ASAuthorizationOpenIDRequest) {
-        nonce = randomNonceString()
+        nonce = cryptoUtils.randomNonceString
         request.requestedScopes = [.fullName, .email]
-        request.nonce = sha256(nonce)
+        request.nonce = cryptoUtils.nonce
     }
     
     func signInWithAppleCompletion(_ result: Result<ASAuthorization, Error>) {
@@ -48,38 +48,5 @@ final class AppleSignInHelper: NSObject, ObservableObject {
         case .failure(let failure):
             print(failure.localizedDescription)
         }
-    }
-    
-    //난수 생성기
-    private func randomNonceString(length: Int = 32) -> String {
-        precondition(length > 0)
-        var randomBytes = [UInt8](repeating: 0, count: length)
-        let errorCode = SecRandomCopyBytes(kSecRandomDefault, randomBytes.count, &randomBytes)
-        
-        if errorCode != errSecSuccess {
-            fatalError(
-                "Unable to generate nonce. SecRandomCopyBytes failed with OSStatus \(errorCode)"
-            )
-        }
-        
-        let charset: [Character] =
-        Array("0123456789ABCDEFGHIJKLMNOPQRSTUVXYZabcdefghijklmnopqrstuvwxyz-._")
-        
-        let nonce = randomBytes.map { byte in
-            // Pick a random character from the set, wrapping around if needed.
-            charset[Int(byte) % charset.count]
-        }
-        
-        return String(nonce)
-    }
-    
-    private func sha256(_ input: String) -> String {
-        let inputData = Data(input.utf8)
-        let hashedData = SHA256.hash(data: inputData)
-        let hashString = hashedData.compactMap {
-            String(format: "%02x", $0)
-        }.joined()
-        
-        return hashString
     }
 }
